requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
networkx>=3.0
pandas>=2.0
numpy>=1.24
requests>=2.31
python-dotenv>=1.0
scipy>=1.10
kiwipiepy>=0.18
wordcloud>=1.9
matplotlib>=3.7
scikit-learn>=1.3
